tt: replace "fork and exec" with "exec" to restart "tt"
For "tt", the standard situation is when the system "tt" after
launch is replaced with "tt" from the environment (local).
It's better to use "exec" instead of "RunExec" to restart "tt"
for this target because we have no reason to have several "tt"
processes. Moreover, it looks strange (and makes harder to stop)
when we start "tt", which starts "tt", which starts tarantool
or some external module.