tt: replace `args[0]` with absolute path to check the using binary
For "tt", the standard situation is when the system "tt" after
launch is replaced with "tt" from the environment (local).
To decide whether to replace the used "tt", the currently running "tt"
is compared with the one that should be started according to the
environment and "tarantool.yaml".
Before this patch, `args[0]` was used for this check, and we may fail
because we can compare the absolute path of the "tt" that should be
started with the relative path that was used when the instance was started
(`/home/vasya/tt/tt` with `./tt`), and as a result the same binary `tt`
will be restarted for nothing.