launcher: add box.cfg() wrapper
If tarantool version is above 2.8.1 we use env variables to set
wal_dir, vinyl_dir and memtx_dir,
else we use cfg.wrapper to do it.

This commit solves problem 30 and allows to set
memtx_dir vinyl_dir and wal_dir.
If they are blank it creates datadir in the run folder.

Closes #30