tt: cli language refactoring
Refactoring of cli language that describe
usage of tt commands.
Now a single notation is used to specify
positional arguments through brackets `<>`,
and 3 points `...` is also used to indicate
the possibility of transmitting multiple
parameters.
Fixed inaccuracies in the description of
some parameters that must be not optional,
but required.

Close #63