rocks: fix make command
When the spec file contains a tarantool dependency, luarocks
checks the tarantool version against the _TARANTOOL global variable.
Since luarocks ran inside the tarantool in tarantoolctl, there were
no problems with this. Luarocks launched from tt is served by gopher-lua
and has no direct access to tarantool variables, accordingly,
a new way of checking dependencies was added, namely:

Added variables TT_CLI_TARANTOOL_PATH and TT_CLI_TARANTOOL_VERSION.

This will allow luarocks to correctly add tarantool to the list of
dependencies and run the correct version of tarantool for internal needs.

Also in this patch are outlines for solving the problem with the correct
determination of the path to the directory with the header files of the
tarantool.

`tt rocks make` test was added.

Also, to solve this problem, a commit was made in luarocks:
https://github.com/tarantool/luarocks/pull/12

Closes #114