create: fix template file name rendering
Fix an issue with rendering file name for the files, which names end
with ".tt.template". If a file name is in .tt.template format, it is
renamed to the . The issue is: if file name also contains template
variables like {{.var}} the instantiation engine tries to instantiate
old file name (.tt.template) and rename it. But this file does not
exists after first renaming. So there is "no such file or directory
error". The fix is to update file name after the first renaming.