test: fix daemon unit test
Unit test failed for systemd because of check if
process is daemonized. Daemonization was checked
with PPID == 1. But PPID == 1 is no guarantee
that it is a daemon, for systemd daemon PPID is
systemd unit that managing daemon process.
But if proceess is detached from it's controlling
terminal, it is a daemon. But it is difficult to
check by means of Golang. This commit removes
daemonization check.

Closes #230