dependency: update containerd version to 1.6.12
Vulnerable versions: < 1.5.16

Impact:
A bug was found in containerd's CRI implementation where a user can
exhaust memory on the host. In the CRI stream server, a goroutine
is launched to handle terminal resize events if a TTY is requested.
If the user's process fails to launch due to, for example, a faulty
command, the goroutine will be stuck waiting to send without a
receiver, resulting in a memory leak. Kubernetes and crictl can both
be configured to use containerd's CRI implementation and the stream
server is used for handling container IO.

CVE ID: CVE-2022-23471
GHSA ID: GHSA-2qjp-425j-52j9

The github.com/opencontainers/runc module has also been updated
as a sub-dependency.
The following vulnerabilities have been fixed:

Vulnerable versions: < 1.1.2

Impact:
A bug was found in runc where runc exec --cap executed processes with
non-empty inheritable Linux process capabilities, creating an atypical
Linux environment and enabling programs with inheritable file
capabilities to elevate those capabilities to the permitted set during
execve(2).

This bug did not affect the container security sandbox as the inheritable
set never contained more capabilities than were included in the container's
bounding set.

CVE ID: CVE-2022-29162
GHSA ID: GHSA-f3fp-gc8g-vw66

Vulnerable versions: < 1.0.3

Impact:
In runc, netlink is used internally as a serialization system for specifying
the relevant container configuration to the C portion of our code (responsible
for the based namespace setup of containers). In all versions of runc prior to
1.0.3, the encoder did not handle the possibility of an integer overflow in the
16-bit length field for the byte array attribute type, meaning that a large
enough malicious byte array attribute could result in the length overflowing and
the attribute contents being parsed as netlink messages for container configuration.

This vulnerability requires the attacker to have some control over the configuration
of the container and would allow the attacker to bypass the namespace restrictions
of the container by simply adding their own netlink payload which disables all
namespaces.

Prior to 9c444070ec7bb83995dbc0185da68284da71c554, in practice it was fairly
difficult to specify an arbitrary-length netlink message with most container
runtimes. The only user-controlled byte array was the namespace paths attributes
which can be specified in runc's config.json, but as far as we can tell no
container runtime gives raw access to that configuration setting -- and having raw
access to that setting would allow the attacker to disable namespace protections
entirely anyway (setting them to /proc/1/ns/... for instance). In addition, each
namespace path is limited to 4096 bytes (with only 7 namespaces supported by runc
at the moment) meaning that even with custom namespace paths it appears an attacker
still cannot shove enough bytes into the netlink bytemsg in order to overflow the
uint16 counter.

However, out of an abundance of caution (given how old this bug is) we decided to
treat it as a potentially exploitable vulnerability with a low severity.
After 9c444070ec7bb83995dbc0185da68284da71c554 (which was not present in any release
of runc prior to the discovery of this bug), all mount paths are included as a giant
netlink message which means that this bug becomes significantly more exploitable in
more reasonable threat scenarios.

The main users impacted are those who allow untrusted images with untrusted
configurations to run on their machines (such as with shared cloud infrastructure),
though as mentioned above it appears this bug was not practically exploitable on
any released version of runc to date.

CVE ID: CVE-2021-43784
GHSA ID: GHSA-v95c-p5hm-xq8f