tt: fix a race condition between tt start and tt stop
With this patch a race condition between tt start and tt stop was fixed.

The race condition:
tt start --watchdog, which is invoked by tt start contains the following main steps:

1. Create PID-file.
2. Start tarantool instance
3. Start signals handling loop

Signals handling loop is required to process interrupt signals and
send them to a child Tarantool instance process. But there is a
race condition here. If the system is under high load, a user
can run tt stop between:

1 and 2 steps. This causes tt watchdog to be killed, but PID file
is not removed and contains PID of the dead tt process.
2 and 3 steps. This causes Tarantool instance hanging, because
interrupt signal is not passed to it by watchdog tt:

Closes #264