ee: add a structure for handling ee bundles information
This patch is needed for having the possibility to get
the information not only about the version of SDK bundle,
but also about its source. Otherwise, it is impossible to get
this information from the name of bundle while installing.
Also, there is need to get the information about gc suffix
of ee SDK bundles.

Part of #281