connect: fix change language race in tests
Connect integration tests with change languages workflow:

1. Start tarantool with app script.
2. Connect to started instance with -l option.

There is a race between console socket creation and box.cfg call. The
second step may fail with the following error:

    error: Unable to set language to 'sql' with unconfigured box

because the socket is already created, but box.cfg from application
script is not called yet.

The patch creates an additional file `configured` after `box.cfg({})`
to ensure that is is already complited.

Closes #313