running: avoid starting an already running instance
Before this patch, starting an already running instance did not stop
the existing process. As a result, the PID file of the existing
process was being overwritten by the new start command. This led
to the situation where it was impossible to control the old process,
as the new one was already started. Additionally, if the old
instance had allocated a port, the new process was unable to allocate
it again and failed without the possibility of reallocation without
manual killing of the existing process.

This patch fixes this behavior and avoid starting the new instance
while it is already running.

Closes #451