cluster: add tt cluster command
* `tt cluster show` to show a configuration.
* `tt cluster publish` to publish a configuration.

@TarantoolBot document
Title: new module - `cluster`

The patch introduces a new module `cluster`. The module has two
commands:

* show - to show a configuration for a cluster or an instance.
* publish - to publish a cluster or an instance configuration.

The target could be a cluster or an instance configuration. It could
be a fully-merged (from all sources) configuration for an application
or just a remote configuration part in case of etcd.

Usage:

* `tt cluster show (<APP_NAME> | <APP_NAME:INSTANCE_NAME> | <URI>) [--validate]`
  * The optional `--validate` flag enables validation for a showed
    configuration.
* `tt cluster publish (<APP_NAME> | <APP_NAME:INSTANCE_NAME> | <URI>) file [--force]`
  * The optional `--force` flag skips validation of uploaded
    configuration.

The `APP_NAME` is a name of an application.

The `INSTANCE_NAME` is a name of an instance in the application.

The `URI` specifies a etcd connection settings in the following format:

`http(s)://[username:password@]host:port[/prefix][?arguments]`

* `prefix` - a base path to Tarantool configuration in etcd.

Possible arguments:

* `name` - a name of an instance in the cluster configuration.
* `timeout` - a request timeout in seconds (default 3.0).
* `ssl_key_file` - a path to a private SSL key file.
* `ssl_cert_file` - a path to an SSL certificate file.
* `ssl_ca_file` - a path to a trusted certificate authorities (CA)
  file.
* `ssl_ca_path` - a path to a trusted certificate authorities (CA)
  directory.
* `verify_host` - set off (default true) verification of the
  certificate’s name against the host.
* `verify_peer` - set off (default true) verification of the peer’s
  SSL certificate.

Part of #506