pack: scaffolding for integrity check file generation
This patch provides a noop `integrity` module intended to be used as a
scaffolding for generating integrity check files. The logic of `tt pack`
command is modified to perform calls to appropriate `integrity` methods
wherever it is required. It is designed as a separate module to allow
easy replacement via directive in `go.mod` file.