clean: fix attempt to remove files multiple times
Before this patch, `tt clean` could list files multiple times since it
is possible that memtx_dir, wal_dir, vinyl_dir and log_dir are the same
directory (and the first three are the same by default). It also had
tried to removed them several times, which had resulted in `[ERR]`
message in logs. This patch fixes the issue.

Closes #735