install: fix update of tarantool master
It was not possible to update tarantool using `tt install tarantool <branch_name>`
because `tt` checked that a tarantool's binary file existed for a branch and
rejected to reinstall if so.

After the patch `tt install` updates a tarantool if commit hashes of the
installed tarantool and a last commit on the branch are different.

Closes #690