install: fix update of tt master
It was not possible to update tt using `tt install tt <branch_name>` checked
that a binary file existed for a branch and rejected to reinstall if so.

After the patch `tt install` updates a tt if commit hashes of the
installed tt and a last commit on the branch are different.

Closes #854