cluster: disable selecting key for the patching
It was inconvenient to select a key every time we want to patch config for
some reason using command `tt cluster replicaset` if there is only one
key.

After the patch `tt` skips selecting of key if it is only one key for
choose. If it is not one key, `tt` will print the path to the
destination path in config.

Closes #923