stop:added confirmation
Since the stop command is dangerous and can be harmful
if executed accidentally, it was necessary to add a user
confirmation for its execution to improve security.
To use auto-confirm, you must pass the -y/--yes option,
like this tt stop -y  or you can use --no-prompt
option to skip interactions.