tests: fix rocks install no tarantool in path
There was a problem with rocks installation using `tt rocks install` in
test where tarantool doesn't exist in PATH. It was `mysql` rock which
had problems with building on system.

After the patch rock was replaced on `psql` (PostreSQL connector) due
to we test installation without `tarantool` in PATH and it is doesn't
matter which package we install.

Part of #931