tests: fix instance logger
There was a problem in `TestInstanceLogger` test with the fact that
cleanup after the test recieves error `io: read/write on closed pipe`.

After the patch the order of pipes closing was updated by moving it to
the cleanup block. A check was also added to send a SIGKILL signal to
the process if it is alive, to avoid the `sending a signal to a killed
process` error.

Part of #931