tests: fix running env variables
After updating tarantool sdk version to the 3.2.0 stopping instance in
`test_running_env_variables` causes an error with cleaning all logs in
json format that was written before. So it wasn't possible to determine
json format after stopping instance.

After the patch the order of distructing tarantool was moved to the
special `finally` block. Determination of json format was also updated
with `wait_string_in_file` function with special timeout.

Part of #931