ci: fix connection refused problem
There was a problem on macOS pipeline that integration tests written on
Go fail sometimes with an `connection refused` error. An instance of
Tarantool was unable to start with the specified configuration.

After the patch start options such as `WaitStart` and `RetryTimeout`
were edited to have enough time for Tarantool startup.Also the port
number in replicaset integration test was changed due to collision
with instance from previous integration test `connector`.

Part of #912