`tt replicaset`: add subcommand downgrade
Closes #968

@TarantoolBot document
Title: `tt replicaset downgrade` downgrades database schema.

The `tt replicaset downgrade` command allows for a automate downgrade of each
replicaset in a Tarantool cluster. The process is performed sequentially on
the master instance and its replicas to ensure data consistency. Below are
the steps involved:

For Each Replicaset:
  - **On the Master Instance**:
    1. Run the following commands in sequence to downgrade the schema and take
       a snapshot:
       ```lua
       box.schema.downgrade(<..version..>)
       box.snapshot()
       ```

  - **On Each Replica**:
    1. Wait for the replica to apply all transactions produced by the
       `box.schema.downgrade` command executed on the master. This is done
       by monitoring the vector clocks (vclock) to ensure synchronization.
    2. Once the repica has caught up, run the following command to take
       a snapshot:
       ```lua
       box.snapshot()
       ```

> **Error Handling**: If any errors occur during the downgrade process, the
operation will halt, and an error report will be generated.

---

- Specify the schema version for downgrade

The `tt replicaset downgrade` command requires specifying the target version for
the schema downgrade. This version should be provided using the `--version`
(or `-v`) option. The version must follow the `x.x.x` format, where `x` represents
a numerical value.

To view the list of available downgrade versions, execute
the following command in Tarantool:

```lua
box.schema.downgrade_versions()
```

**Example:**
```bash
$ tt replicaset downgrade [<APP_NAME> | <URI>] --version 3.0.0
```

- Timeout for Synchronization

Replicas will wait for synchronization for a maximum of `Timeout` seconds.
The default timeout is set to 5 seconds, but this can be adjusted manually
using the `--timeout` option.

**Example:**
```bash
$ tt replicaset downgrade [<APP_NAME> | <URI>] -v 3.0.0 --timeout 10
```

- Selecting Replicasets for Downgrade

You can specify which replicaset(s) to downgrade by using the `--replicaset`
or `-r` option to target specific replicaset names.

**Example:**
```bash
$ tt replicaset downgrade [<APP_NAME> | <URI>] -v 3.0.0 replicaset <RS_NAME_1> -r <RS_NAME_2> ...
```

This provides flexibility in downgrading only the desired parts of the cluster
without affecting the entire system.