pack: support .packignore
In some cases, there are files that may be useful, but should not
be part of artifact. The ability to add files and directories to
the .packignore file has been added, which allows you to ignore
these files and directories when packing.

Closes #812

@TarantoolBot document
Title: `tt pack` support `.packignore`

Use `.packignore` in the same way as `.gitignore` allows to exclude
unnecessary files while preparing application package with `tt pack
command.