search: add search TCM in customer zone
1. Added `tt search tcm` command to searching for tarantool cluster
manager (TCM) in the customer zone or local directory.
It displays the available versions in a user-friendly format.

2. Refactoring search - split code into separate files for each
searching types: [git, local, sdk].

@TarantoolBot
Title: added `tt search tcm`

The command performs a search for tarantool cluster manager (TCM)
in the customer zone or local directory configured as `distfiles` in
the tt.yaml. It displays the available versions in a user-friendly format.

**Usage Example**:
```bash
tt search tcm
```

**Output Example**:
```
   • Available versions of tcm:
1.2.0-11-g2d0a0f495
1.2.1-0-gc2199e13e
1.2.3-0-geae7e7d49
1.3.0-0-g3857712a
1.3.1-0-g074b5ffa
```

Closes #TNTP-1094