tcm: show TCM logs
Added support for showing TCM logs

@TarantoolBot
Title: Added support for showing TCM logs

This commit introduces the ability to view TCM logs using
the `tt tcm log` command.

This command works in couple with the `tt tcm start` command,
and show the logs of the TCM process that was running by it.

Command expect exists `tcm.log` file in the current directory.

**Usage example:**
```bash
tt tcm log --lines 100 --follow
```

This command will show the last 100 lines of the TCM log
file and continue to follow the log output.
This feature is useful for monitoring the TCM process.
Without the `--follow` flag, it will show the last 100 lines
of the log file and exit.

The command also supports following flags:
- `--no-color` to disable colored output
- `--no-format` to disable structured output

**Possible output:**
```text
{
  time: **2025-05-29T17:19:47.905335965+03:00**
  level: INFO
  msg: _tcm stopped_
  source: {
    file: "tcm/internal/app/tcm/server/server.go",
    function: "tcm/internal/server.(*TcmAPI).Close",
    line: 175
  }
}
{
  time: **2025-05-29T17:19:47.90451709+03:00**
  level: WARN
  msg: _retrying of unary invoker failed_
  attempt: 0,
  error: "rpc error: code = Canceled desc = latest balancer error:
    last connection error: connection error: desc = \"transport:
    Error while dialing: dial tcp 127.0.0.1:2379: connect:
    connection refused\"",
  source: {
    file: "go.etcd.io/etcd/client/v3@v3.5.15/retry_interceptor.go",
    function: "v3.(*Client).unaryClientInterceptor.func1",
    line: 63
  },
  target: "etcd-endpoints://0xc0003d2000/127.0.0.1:2379"
}
{
  time: **2025-05-29T17:19:47.90548212+03:00**
  level: ERROR
  msg: _fail to get entity from etcd_
  err: "context canceled",
  path: "/cluster/",
  source: {
    file: "tcm/internal/app/tcm/server.go",
    function: "tcm/internal/app/tcm/server.(*TcmAPI).validationKV",
    line: 17
  }
}
```

Closes #TNTP-1103