logrotate: fix log rotation
Watchdog launches tarantool in a way that produce log to stderr and thus
tarantool itself does nothing to handle log rotation, it is handled
entirely at watchdog side.
Prior to this patch `os.File` was used in `Cmd.Stdout`, when watchdog
launches tarantool binary and according to golang documentation
(see `Exec.Cmd.Stdout`) in this special case output from the child
process is connected directly to that file. It means that even if it is
closed at watchdog side, its child process (tarantool) keeps using the
original one, i.e. it proceeds with logging to the original file even
after renaming it.
Now `Logger` is used as `Cmd.Stdout`, and according to the mentioned
documentation separate goroutine reads output from the child process
over a pipe and delivers that data to the `Logger`. It allows proper
handling of log rotation at watchdog side.

Closes #1174

@TarantoolBot document
Title: `tt logrotate` properly release descriptor of the old log file.