switch: minor adjustment of command output
At the end of its output `switch` displays command `switch-status` with
the corresponding arguments that allows to check status of switching.
This patch wraps uri argument of that command with quotes as it may
contain characters that have special meaning for shell, thus it simplify
subsequent call of `switch-status` command because now it can be
copy-pasted from the output as-is.

Closes #1194

@TarantoolBot document
Title: `tt cluster failover switch` simplify subsequent launch of
`switch-status`.