cd: add CentOS 8
Similarly to commit 84850608a5c2 ("cd: add Ubuntu Noble") it is
sometimes convenient to have the `tt` package right in the distro
specific repository even if it is available in the general
`linux-rpm/static` repository.

Since it costs nothing (the same RPM file is published to different
repositories), I propose to publish the `tt` package to the `el/8`
repository as well.

See tarantool/installer.sh#42 for the story behind it.

Fixes #1192