cluster: fix 'no-ssl' build
follow-up 331436c2 applying the approach introduced in 97f88ece

The mentioned approach implies using of `lib/dial` library rather than
`tls-dialer`. The implementation of `lib/dial` depends on build type
and either uses `tls-dialer` for 'ssl' build (`tls-dialer` in turn
depends on OpenSSL c-library) or dummy implementation for 'no-ssl'.

Prior to this patch `tls-dialer` was used directly which means that
`tt`-executable always required OpenSSL c-library (`libssl.so`).

Certain version of the SSL-library is resolved at build-time and depends
on the build environment. So previously `tt`-executable always required
`libssl.so` of the version that was installed on builder. Then if target
system didn't have it `tt` executable failed to launch. This exact
problem arose for example when one tries to install `tarantool` using
docker with command:

$ tt install tarantool --use-docker

This command eventually launch `tt` in a docker image that is prepared
internally and is based on Ubuntu 18.04 which have SSL of version 1
onboard. `tt` itself is usually built in modern environment with SSL of
version 3 and thus its execution in docker failed.

Closes #1220
Part of #1223