test: use predefined versions of tarantool
Prior to this patch integration test `test_install_tarantool_in_docker`
checked doesn't specify tarantool version explicitly, so it was always
checked against the latest one. Which means that it might fail due to
some problem introduced in tarantool and thus broke `tt` release cycle
with non-tt problem.

To avoid such an unpredictable behavior this test is adjusted to check
only against predefined `tarantool` versions (each of them is obtained
from the corresponding test environment).
Original test that checks against the latest `tarantool` version is to
be moved to separate workflow that is not involved into release cycle
and is performed periodically (by cron) to monitor consistency.

Closes #1221