docker: bump tarantool builder to Ubuntu 20.04
`tt` performs `tarantool` in-docker compilation in the scope of commands

tt install tarantool --use-docker
tt pack tgz|deb|rpm --use-docker

Docker images that are used internally are based on Ubuntu 18.04, but
starting from tarantool/tarantool#1fd4404 `tarantool` compilation no
longer succeed in this system due to compilation issues. The problem
has been fixed already (tarantool/tarantool#12170), but anyway there are
a bunch of commits (including version 3.6.0 of tarantool) that can't be
used with the commands mentioned above.

At the same time Ubuntu 18.04 is quite an old release that is no longer
supported. So this patch bumps "base" of the tarantool-builder dockers
to the next Ubuntu LTS which is 20.04 (its default c-compiler is not
affected with the "problem" commits).

Closes #1223
Related to tarantool/tarantool#12167